requirements.txt
# fastapi
# uvicorn
# pymongo
# chromadb
# python-dotenv
# pydantic
# groq
# reportlab
# onnxruntime

fastapi
uvicorn
pymongo
python-dotenv
pydantic
groq
reportlab